chore: move developer-options-tab into settings-v2/debug-tab
The legacy developer-options-tab folder under ui/pages/settings/ was
re-exported wholesale by ui/pages/settings-v2/debug-tab/debug-tab.tsx
(its only consumer). Co-locating the implementation with its V2 wrapper
removes the cross-folder dependency and lets us close out
ui/pages/settings/ piece-by-piece without leaving dangling re-exports.

- git-rename ui/pages/settings/developer-options-tab/ ->
  ui/pages/settings-v2/debug-tab/developer-options-tab/ (12 files,
  including the snapshot).
- Bump every relative import in the moved files by one extra `../` so
  selectors/store-actions/component-library/etc. still resolve from the
  deeper location.
- Update the import in settings-v2/debug-tab/debug-tab.tsx to point at
  ./developer-options-tab/developer-options-tab.
- Repoint the @import line in ui/pages/settings/index.scss at the new
  location so the .settings-page-developer-options rule keeps loading
  (still consumed by ui/pages/confirmations/.../developer-button).
- Update the two console-baseline-unit.json paths.

No production behaviour change; the rendered tree, route, and CSS class
names are unchanged.

Made-with: Cursor
