+++ test/jest/console-baseline-unit.json
@@ -1590,11 +1590,11 @@
       "Reselect: Identity function warnings": 1,
       "error: Warning: Received `true` for a": 1
     },
-    "ui/pages/settings/developer-options-tab/developer-options-tab.test.tsx": {
+    "ui/pages/settings-v2/debug-tab/developer-options-tab/developer-options-tab.test.tsx": {
       "React: Act warnings (component updates not wrapped)": 11,
       "React: componentWill* lifecycle deprecations": 1
     },
-    "ui/pages/settings/developer-options-tab/developer-options-toggle-row-component.test.tsx": {
+    "ui/pages/settings-v2/debug-tab/developer-options-tab/developer-options-toggle-row-component.test.tsx": {
       "React: componentWill* lifecycle deprecations": 1
     },
     "ui/pages/settings/networks-tab/networks-form/networks-form.test.js": {
