+++ ui/pages/settings-v2/debug-tab/debug-tab.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { SettingsTab } from '../shared';
 import { SettingItemConfig } from '../types';
-import DebugTabContent from '../../settings/developer-options-tab/developer-options-tab';
+import DebugTabContent from './developer-options-tab/developer-options-tab';
 
 const DEBUG_SETTING_ITEMS: SettingItemConfig[] = [
   { id: 'debug', component: DebugTabContent },
