+++ ui/pages/settings-v2/debug-tab/developer-options-tab/developer-options-toggle-row-component.test.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import configureMockStore from 'redux-mock-store';
 import thunk from 'redux-thunk';
-import { renderWithProvider } from '../../../../test/lib/render-helpers-navigate';
-import mockState from '../../../../test/data/mock-state.json'; // Adjust the path based on your actual file location
+import { renderWithProvider } from '../../../../../test/lib/render-helpers-navigate';
+import mockState from '../../../../../test/data/mock-state.json'; // Adjust the path based on your actual file location
 import ToggleRow from './developer-options-toggle-row-component';
 
 describe('ToggleFeatureRow Component', () => {
