+++ ui/pages/settings-v2/debug-tab/developer-options-tab/index.scss
