+++ ui/pages/settings/index.scss
@@ -1,6 +1,6 @@
 @use "design-system";
 
-@import 'developer-options-tab/index';
+@import '../settings-v2/debug-tab/developer-options-tab/index';
 @import 'networks-tab/index';
 @import 'security-tab/index';
 @import 'transaction-shield-tab/index';
